Add mistral-ceph-hci deployemtn THT and Script
- deploy-mistral-ceph-hci.sh will deploy a two-node OpenStack
  to use an extenral ceph cluster
- ironic.sh was updated to tag nodes for this new deployment

The idea is to use this and the run an independent mistral workflow
to deploy Ceph on the same nodes:

 https://github.com/fultonj/mistral/tree/master/mistral-ceph-ansible

From there it's just a matter of restarting the OpenStack services and
this was tested to work with OSP8+Ceph1.3:

 https://github.com/RHsyseng/conv-osp-ceph/blob/osp8_ceph1.3/scripts/connect_osp_ceph.sh

Though this type of deploy worked in the past but is currently running into
an issue with the hash the is generated by THT being fed to puppet-ceph
as a string:

Error: Evaluation Error: Error while evaluating a Function Call, "{ client.openstack: { secret: 'AQBtcrVWAAAAABAA4JprB4BR5klHr3mgtqmX0w==', mode: '0644', cap_mon: 'allow r', cap_osd: 'allow class-read object_prefix rbd_children, allow rwx pool=volumes, allow rwx pool=backups, allow rwx pool=vms, allow rwx pool=images, allow rwx pool=metrics' } }" is not a Hash.  It looks to be a String at /etc/puppet/modules/ceph/manifests/profile/params.pp:243:3 on node overcloud-novacompute-0.localdomain

This is because this call receives a string not a hash:

 https://github.com/openstack/puppet-ceph/blob/master/manifests/profile/params.pp#L239

The string is created here:

 https://github.com/openstack/tripleo-heat-templates/blob/6b8c34b27a324120f6945f7873f479d37ba2e516/puppet/services/ceph-external.yaml#L84-L101

So this almost works, just want to workaround the above.